Added a CodeMirror html file
Created a code mirror html file to practice adding a code editor on the
codeWarrior exercises. The CodeMirror files are being hosted through
google drive.

To host the pages, we need to upload the specific code-mirror files into
google drive and get a shareable link that can be accessed by anyone.
Then we need to isolate the link id (example:
1D3aUBXId9Br8mhNhoXY0n3hAHQkXONoQ) and added on at the end of the
following url:
https://drive.google.com/uc?export=view&id=FIILE-ID-GOES-HERE .Finally,
all we need to do is include the new url link onto a script tag if it is
a js file or a css link tag if the file is css.